Update Navier-Stokes for recent changes
